+++ clfd/__init__.py
@@ -1,3 +1,7 @@
+# NOTE: this must be imported first
+from ._version import __version__
 from .core import DataCube, featurize, profile_mask, time_phase_mask
+from .report import Report
+from .tests import test
 
-__all__ = ['DataCube', 'featurize', 'profile_mask', 'time_phase_mask']
+__all__ = ['DataCube', 'featurize', 'profile_mask', 'time_phase_mask', 'Report', 'test']