+++ clfd/tests/__init__.py
@@ -1,0 +1,3 @@
+from .run_tests import test
+
+__all__ = ['test']