+++ clfd/tests/test_data_cube.py
@@ -21,7 +21,11 @@ class TestDataCube(unittest.TestCase):
         self.ndarray = numpy.load(self.npy_data_fname)
 
     def test_init(self):
-        clfd.DataCube(self.ndarray)
+        # NOTE: need to leave self.ndarray untouched. It WILL be modified
+        # if wrapping it in a DataCube with copy=False
+        X = self.ndarray.copy()
+        clfd.DataCube(X, copy=False)
+        clfd.DataCube(X, copy=True)
 
     def test_input_type_errors(self):
         # Input must be numpy.ndarray
@@ -44,21 +48,20 @@ class TestDataCube(unittest.TestCase):
             clfd.DataCube(self.ndarray[:1, :1, :])
 
     def test_load_save_npy(self):
-        cube1 = clfd.DataCube(self.ndarray)
         with tempfile.NamedTemporaryFile(mode="wb", suffix='.npy', delete=False) as fobj:
             fname = fobj.name
         
-        cube1 = clfd.DataCube(self.ndarray)
+        # NOTE: leaving self.ndarray untouched
+        cube1 = clfd.DataCube(self.ndarray, copy=True)
         cube1.save_npy(fname)
         cube2 = clfd.DataCube.from_npy(fname)
+
         self.assertTrue(numpy.allclose(cube1.data, cube2.data))
         os.remove(fname)
 
     @unittest.skipUnless(HAS_PSRCHIVE, "psrchive python bindings must be installed")
     def test_load_psrchive(self):
         cube = clfd.DataCube.from_psrchive(self.psrchive_data_fname)
-
-
 
 
 if __name__ == "__main__":